requirements.txt
pandas
tqdm
simpletransformers
lm-scorer
nemo_toolkit[all]==0.11.0
torch==1.5.1
torchaudio==0.5.1
torchvision==0.6.0
transformers==2.11.0
tokenizers==0.7.0